requirements.txt
PyPDF2==3.0.1
sentence-transformers 
faiss-cpu==1.9.0.post1
streamlit==1.41.1
numpy==2.2.1
PyMuPDF==1.25.1 